fix(KEN-821): one reading of what a pinned revision supplies
The inventory of skills a scope can supply read the manifest's pins
directly, while the planner answered the same question through the
expansion, the set expansion and the dependency walk. The two disagreed:
a pinned skill's dependencies install from the parent's own commit, and
the inventory counted only the skill named in the pin, so a dependency
the source has since dropped was called absent — the agent assigned it
rendered without the row, and a fork of that agent was refused over a
skill sitting in the scope's own tree.

Pinned availability now comes from the closure the planner already
derives: planned skills carrying a revision, confirmed present at it.
A skill declared at a pin, a set member the set expansion kept and a
dependency read from its parent's commit all arrive the same way, so
the set-member and lands-nowhere checks the inventory used to spell for
itself are the planner's again.

ScopeSkills moves from source to engine, where the closure and all four
of its callers live. What a source offers at its own revision is
unchanged: an assignment naming an offered skill is answered by
declaring it, and narrowing that to installed-only refuses a fork the
scope can answer.

fork and detach gain an expansion each to ask the question; desired
passes the one it already has.
