+++ changelog.d/fixed/KEN-821.md
@@ -1,0 +1,3 @@
+- A skill only a pinned revision carries — a dependency of a pinned package —
+  is available again: the agent assigned it renders the row, and forking that
+  agent is no longer refused.
